fix: price history shows real multi-point data from RentCast raw_data
Root cause: property_service.get_detail() never read the 'history' field
already present in every RentCast listing's raw_data. It fell through to
synthesizing a single point from current_price, giving the chart only one
data point no matter what.

Changes:
- get_detail() now reads raw_data['history'] (YYYY-MM-DD keyed dict with
  price, listedDate, daysOnMarket per entry) to build real price events;
  ATTOM is still tried for tax history but price history comes from here
- synthetic_only check: if existing DB history is just one 'listing'-source
  event (old synthetic), it's treated as missing so raw_data is extracted
- _persist_history(replace_price=True) deletes the old synthetic row before
  inserting real events, preventing stale state on re-views
- expire_all() called before the post-persist re-fetch so SQLAlchemy doesn't
  return identity-map cached data from within the same request
- PriceEvent.price and TaxRecord.assessed_value/tax_amount changed from
  Decimal to float in schema — Pydantic v2 was serializing them as strings,
  which would silently break the chart's number parsing
- Cleared 13 old synthetic 'listing'-source events from price_history so
  all properties get real history on their next detail page view

+++ backend/app/schemas/property.py
@@ -7,7 +7,7 @@ from decimal import Decimal
 class PriceEvent(BaseModel):
     id: UUID4
     event_type: str
-    price: Decimal
+    price: float
     event_date: date
     source: str
     notes: Optional[str] = None
@@ -18,8 +18,8 @@ class PriceEvent(BaseModel):
 class TaxRecord(BaseModel):
     id: UUID4
     year: int
-    assessed_value: Optional[Decimal] = None
-    tax_amount: Optional[Decimal] = None
+    assessed_value: Optional[float] = None
+    tax_amount: Optional[float] = None
     source: str
 
     model_config = {"from_attributes": True}
